[Fix]  reference: tap하여 삭제버튼 표시

+++ QueenCam/QueenCam/Presentation/Views/Reference/OpenView.swift
@@ -30,9 +30,10 @@ struct OpenView: View {
         }
       }
 
-      if showDelete || (referenceViewModel.image != nil) {
+      if showDelete && (referenceViewModel.image != nil) {
         Button {
           referenceViewModel.onDelete()
+          showDelete = false
         } label: {
           Image(systemName: "x.circle")
             .imageScale(.large)
